requirements.txt
python-telegram-bot
requests==2.13.0
backports.functools_lru_cache==1.5
enum34
configparser